[Misc] Restructure backends, fix issues with build
