+++ Xenon/Core/XGPU/CommandProcessor.h
@@ -16,8 +16,9 @@
 #pragma once
 
 #include <atomic>
-#include <thread>
+#include <condition_variable>
 #include <memory>
+#include <thread>
 #include <unordered_map>
 
 #include "Base/Logging/Log.h"
